+++ src/index.ts
@@ -1,7 +1,5 @@
 export * from './lib/permissions/index';
 export * from './lib/user/index';
-
-export * from './lib/testing/index';
 export * from './shared';
 
 export * from './lib/pages/index';
