+++ src/lib-methods.ts
@@ -1,0 +1,8 @@
+export {
+  MultiPageTreeResolveInput,
+  MultiPageTreeResolveOutput,
+  multiResolvePageTree,
+  resolvePageTree
+} from './lib/pages/resolvePageTree';
+
+export {} from './lib/permissions/hasAccessToSpace';
