--- src/lib/forums/posts/interfaces.ts
@@ -1,3 +1,0 @@
-import type { Post, PostUpDownVote } from '@prisma/client';
-
-export type ForumPostWithVotes = Post & { upDownVotes: Pick<PostUpDownVote, 'upvoted' | 'createdBy'> };
