+++ src/lib/permissions/pages/index.ts
@@ -1,4 +1,4 @@
 export * from './policies';
-export * from './availablePagePermissions.class';
+export { AvailablePagePermissions } from './availablePagePermissions.class';
 export * from './interfaces';
 export { copyAllPagePermissions, copyPagePermission } from './copyPagePermissions';
