+++ src/pages.ts
@@ -1,0 +1,2 @@
+export * as PageTypes from './lib/pages/interfaces';
+export * as pageTree from './lib/pages/mapPageTree';
