+++ src/permissions.ts
@@ -1,0 +1,4 @@
+export * as PermissionTypes from './lib/permissions/interfaces';
+
+export { AvailablePagePermissions } from './lib/permissions/pages/availablePagePermissions.class';
+export { PagePermissionsHttpClient } from './lib/permissions/pages/client/pagePermissionsHttpClient';
